Removed casting to int for angle and angle step size.  Also switch of Preview checkbox for fitting plugin after the fit finished

+++ src/main/java/edu/ucsf/valelab/saim/SaimFit.java
@@ -94,8 +94,8 @@ public class SaimFit implements PlugIn, DialogListener {
          sd_.wavelength_ = gd.getNextNumber();
          sd_.nSample_ = gd.getNextNumber();
          sd_.dOx_ = gd.getNextNumber();
-         sd_.firstAngle_ = (int) gd.getNextNumber();
-         sd_.angleStep_ = (int) gd.getNextNumber();
+         sd_.firstAngle_ = gd.getNextNumber();
+         sd_.angleStep_ = gd.getNextNumber();
          sd_.mirrorAround0_ = gd.getNextBoolean();
          sd_.zeroDoubled_ = gd.getNextBoolean();
          sd_.A_ = gd.getNextNumber();
